fix: inject vscode and state into managers
